+++ drift/CHANGELOG.md
@@ -3,6 +3,9 @@
 - Add `isExp`, `isValue`, `isNotExp` and `isNotValue` methods to `Expression`
   to generate the `IS` operator in SQL.
 - Add `all()` extension on tables and views to quickly query all rows.
+- Add `serializableConnection()` and `computeWithDatabase()` as extensions on
+  databases. The methods allow sharing any drift database between isolates,
+  regardless of how the database has been set up.
 - The `DatabaseConnection` class now implements `QueryExecutor`, meaning that
   you no longer need a special `.connect()` constructor to use it.